requirements.txt
Pillow==11.2.1
pyautogui==0.9.54
pydirectinput==1.0.4
pygetwindow==0.0.9
pywinauto==0.6.9
sv_ttk==2.6.0
pywin32==310
opencv-python==4.11.0.86 
